+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Added `unzip` for splitting pairs into two collections.
 - Added `min_by_key` for selecting the first minimum keyed item.
 - Added `max_by_key` for selecting the first maximum keyed item.
+- Added `defaults` for first-value-wins map merging.
 
 ## [0.7.0] 2026-08-02
 ### Added